// src/interface/interfaces.ts
  interface HourlyTemperatureAndHumidity {
    time: string;
    humidity: number;
    temperature: number;
  }
  
  interface pressureSurfaceLevelByTemperature {
    pressureSurfaceLevel: number;
    temperature: number
  }